The proof of finterp_1 is done.
We can simplified the proof as Younesse done for tinterp_1.

I want to say one important point in the proof:
Actually, we are doing induction on A and not on v0, v1 and v2. But we should consider the v0, v1 and v2 as general case (arbitrary v0, v1 and v2) and not specific through the proof. However I only used any arbitrary v0 (not for v1 and v2) for this proof and it helped me to prove the existential and forall quantifier. The tactic to doing that, as you see in the proof, is "revert".